Change MatchData to work better with db.Model
Each db.Column becomes a property that can be used to get/set the MatchData object's state. However, I was keeping the db Columns seperate from the MatchData's attributes, meaning that saving the MatchData state into the db would be difficult when it came time to commit any new subissions.

To fix this, where the db.Column property was `C_TEAM_NUMBER` and the attribute was `team_number`, i just made the db.Column property `team_number` and used that instead of an attribute.